added ingestion_status tool
this tools
(1) Uses file-mask in the ingestion files to check the raw data
repository (day) for availability:
File count > 0 means data files available.
File count = 0 means no files are found: check path or files are
missing.

(2) append the start and end times from the deployment sheet to every
row on the ingestion file. This step is needed to create the timelines.

(3) create a file with a status column to be checked and completed
using the Status_key.csv file found under the tool box.

(4) create a timeline for every platform by reference designator and
delivery method.